fix: array file-scope decay come arg, %X/%llX runtime, uint64_t 64-bit, >> u64
custom_lib/des.c (DES completo) non compilava/girava 1:1. Trovati e risolti:
- array file-scope/static passato per decay (`f(E,…)` con `E` globale): slot
  base cercato anche in mem_layout (`_prepare_call_arg`, `_lower_assign`).
- `printf %X`/`%llX` runtime (era solo costanti): nuove proc lib uppercase
  `__mn_putX*` + `__mn_putX_u64_width_zero` per `%016llX`.
- `uint64_t`/`int64_t`/`uintmax_t` ora 64-bit (fake_include → `long long`) e
  ESCLUSI dal wrap-32 (`_uint32_scalar_kind` salta i `long long`): prima
  `1ULL<<63` veniva mascherato a 0.
- `_expr_is_unsigned` espande i typedef (uint64_t → unsigned) e usa l'operando
  sinistro per gli shift → `u64 >> n` resta logico (non floor-div signed che
  con `1<<64`=0 dava divmod b==0).

+++ mnemo/fake_include/stdint.h
@@ -6,14 +6,14 @@
 typedef int int8_t;
 typedef int int16_t;
 typedef int int32_t;
-typedef int int64_t;
+typedef long long int64_t;
 typedef unsigned int uint8_t;
 typedef unsigned int uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned int uint64_t;
+typedef unsigned long long uint64_t;
 typedef int intptr_t;
 typedef unsigned int uintptr_t;
-typedef int intmax_t;
-typedef unsigned int uintmax_t;
+typedef long long intmax_t;
+typedef unsigned long long uintmax_t;
 
 #endif
